FROMLIST: mmc: core: Support the optional init_card() callback for MMC and SD
In (3fcb027 ARM: MXC: mxcmmc: work around a bug in the SDHC busy line
handling) the optional init_card() callback was added.  According to
the original change it was "for now only called from
mmc_sdio_init_card()".

This callback really ought to be called from the SD and MMC init
functions as well.  One current user of this callback
(mxcmci_init_card) will not work as expected if you insert an SDIO
card, then eject it and put a normal SD card in.  Specifically the
normal SD card will not get to run with 4-bit data.

I'd like to use the init_card() callback to handle a similar quirk on
dw_mmc when using SDIO Interrupts (the "low power" feature of the card
needs to be disabled), so that will add a second user of the function.

BUG=chrome-os-partner:33201
TEST=WiFi speed is faster.

Change-Id: Ice9a166d2c155419a5aeed371b39f342e9361bb7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
(am from https://patchwork.kernel.org/patch/5425231/)
Reviewed-on: https://chromium-review.googlesource.com/232777
Reviewed-by: Alexandru Stan <amstan@chromium.org>