UPSTREAM: mmc: sdhci: Tuning should not change max_blk_count
Re-tuning requires that the maximum data length
is limited to 4MiB. The code currently changes
max_blk_count in an attempt to achieve that.
This is wrong because max_blk_count is a different
limit, but it is also un-necessary because
max_req_size is 512KiB anyway. Consequently, the
changes to max_blk_count are removed and the
comment for max_req_size adjusted accordingly.
The comment is also tweaked to show that the 512KiB
limit is a SDMA limit not an ADMA limit.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Reviewed-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit ac00531d9f4be487366573e2bb943e68e9e523ab)
Signed-off-by: Marc Herbert <marc.herbert@intel.com>

BUG=chrome-os-partner:40410
TEST=With this whole HS400 series applied:
TEST=on a system with HS400 controller(s), it/they initialize with the
  following log message:
  INFO kernel: [  103.601747] mmc0: new HS400 MMC card at address 0001
TEST=run SD/eMMC test plan: higher eMMC throughput after change, and no
  other change.
TEST=same on another system with no HS400: no observable change at all.

Change-Id: I48d3723311c2c5c3e64078a036ad4ca92adbadf8
Reviewed-on: https://chromium-review.googlesource.com/276390
Commit-Ready: Marc Herbert <marc.herbert@intel.com>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>