drm/tegra: dc: Initialize border color
Tegra114 and earlier support specifying the color of the border (i.e.
the active area of the screen that is not covered by any of the overlay
windows). By default this is set to a light blue, so set it to black to
comply with the requirements set by atomic modesetting.

Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 42d0659ba76956665c704d2a7cee9fa0f54a7acf)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>