UPSTREAM: rhashtable: Convert bucket iterators to take table and index
This patch is in preparation to introduce per bucket spinlocks. It
extends all iterator macros to take the bucket table and bucket
index. It also introduces a new rht_dereference_bucket() to
handle protected accesses to buckets.

It introduces a barrier() to the RCU iterators to the prevent
the compiler from caching the first element.

The lockdep verifier is introduced as stub which always succeeds
and properly implement in the next patch when the locks are
introduced.

Signed-off-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 88d6ed15acff1cb44b1d1f3c0a393b7f7744957a)

BUG=chromium:556861
TEST=builds all four kernel flavors based on chromeos-3.18 branch

Change-Id: I28cea8236f2abf2bc8a1af263675c0fea541add0
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/314141
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>