FROMLIST: mfd: Add support for Intel Sunrisepoint LPSS devices
The new coming Intel platforms such as Skylake will contain Sunrisepoint PCH.
The main difference to the previous platforms is that the LPSS devices are
compound devices where usually main (SPI, HSUART, or I2C) and DMA IPs are
present.

This patch brings the driver for such devices found on Sunrisepoint PCH.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://patchwork.kernel.org/patch/6607801/
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

[Vedang Patel]: Using clkdev_alloc + clkdev_add instead of clkdev_create
because the helper was added in v4.1 and is not currently there in
chromeos-3.18.

Conflicts:
	(Minor differences in Makefile)
	drivers/mfd/Makefile

BUG=chrome-os-partner:40857
TEST=In the output of lspci -v, intel-lpss name gets attached to most
   "Signal processing controller" sections.
TEST=Two new lines "Synopsys Designware I2C adapter" appear in:
 - cat /sys/bus/i2c/devices/*/name
 - and in the output "i2cdetect -l" (i2cdetect on "test" image at least)
TEST=Two new "i2c_designware" lines appear in /proc/interrupts

Change-Id: I4766bbc97383abe8a90e8ac5ce9beb5694782714
Reviewed-on: https://chromium-review.googlesource.com/282878
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>