verity: Add error path for ubiblock-backed dm-verity
The typical dm-verity-chromeos error path writes to the kernel partition
using a bio. For raw NAND, the kernel partition is a bare mtd partition.
It cannot be written to directly. This patch updates the error path to
detect if it is running on raw NAND (by checking if the rootfs is
mounted on an ubiblock) and, if so, to translate the block device into
the corresponding kernel partition and stamp the DMVERROR onto it using
mtd commands.

BRANCH=none
BUG=chromium:442794
TEST=Corrupted root-a and booted from kern-a, observing a correct reboot
following that into kern-b, which mounted the verified rootfs and got
to the shell.

Change-Id: Ic1b7e714b7ec02ec6f21f6f8408eb379d8fe428d
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238573
Reviewed-by: Kees Cook <keescook@chromium.org>