UPSTREAM: xhci: Silence "xHCI xhci_drop_endpoint called with disabled ep ..." messages
When re-applying the configuration after a successful usb device reset,
xhci_discover_or_reset_device has already dropped the endpoints, and free-ed
the rings.

The endpoints already being dropped is expected, and should not lead to
warnings. Use the fact that the rings are also free-ed in this scenario to
detect this, and suppress the "xHCI xhci_drop_endpoint called with disabled
ep ..." message in this case.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>

--
Changes in v2:
Move the ring check to only guard the xhci_warn, so as to avoid side-effects
in case we have a scenario where the rings are free-ed, but the endpoint is
not yet dropped.
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
br: usb-next
(cherry picked from commit a6134136d938ed9298f15e865e4a035f9c0eeb9c)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the USB subsystem is more stable.
TEST=there are no regressions or file system related errors observed
TEST=on SKL based systems.

Change-Id: Idf7e0d2eacf1b35a72cb206e1fcd4c9e35d9612c
Reviewed-on: https://chromium-review.googlesource.com/272759
Tested-by: Philip Hanson <philip.hanson@intel.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>